included branches in create and edit employee

+++ LO.CD.Entities/Employee.cs
@@ -21,7 +21,7 @@ namespace LO.CD.Entities
         public string LastName { get; set; }
         public Gender Gender { get; set; }
         public string Image { get; set; }
-        public double Sallary { get; set; }
+        public int Sallary { get; set; }
 
         public List<Deal> Deals { get; set; }
